-1.0 instead of casting to float

+++ azurerm/internal/services/loganalytics/log_analytics_workspace_resource.go
@@ -76,7 +76,7 @@ func resourceArmLogAnalyticsWorkspace() *schema.Resource {
 			"daily_quota_gb": {
 				Type:         schema.TypeFloat,
 				Optional:     true,
-				Default:      float64(-1),
+				Default:      -1.0,
 				ValidateFunc: validation.Any(validation.FloatBetween(-1, -1), validation.FloatAtLeast(0)),
 			},
